requirements.txt
# FastAPI 및 웹 서버
fastapi==0.116.1
uvicorn==0.35.0

# 데이터베이스
SQLAlchemy==2.0.43
aiosqlite==0.21.0
alembic==1.16.4
greenlet==3.0.3

# 인증 및 보안
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==45.0.6

# 데이터 검증
pydantic>=2.7.0
pydantic-settings==2.2.1
email-validator==2.1.0

# HTTP 클라이언트
httpx==0.28.1
requests==2.31.0

# OAuth 라이브러리
google-auth==2.28.1

# Google Sheets API
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0

# 템플릿 엔진 (Alembic용)
Mako==1.3.10

# YAML 설정 파일 지원
PyYAML==6.0.1

# MySQL 드라이버
aiomysql==0.2.0

# 이미지 처리
Pillow==10.4.0

# Sendbird Platform SDK
sendbird-platform-sdk==2.1.0

# Firebase Cloud Messaging (푸시 알림)
firebase-admin==6.5.0

# Monitoring (모니터링)
prometheus-fastapi-instrumentator==7.0.0

